Rename helper files and add CI workflow
