+++ src/components/section-card/__tests__/section-card.test.tsx
@@ -3,11 +3,13 @@ import { SectionCard } from '../section-card';
 
 it('renders title and children', () => {
   render(
-    <SectionCard title="My Title">
+    <SectionCard title='My Title'>
       <p>Hello world</p>
     </SectionCard>
   );
 
-  expect(screen.getByRole('heading', { level: 2 })).toHaveTextContent('My Title');
+  expect(screen.getByRole('heading', { level: 2 })).toHaveTextContent(
+    'My Title'
+  );
   expect(screen.getByText('Hello world')).toBeInTheDocument();
 });
