+++ src/helpers/parse-highlight-lines.ts
@@ -6,7 +6,9 @@ export const parseHighlightLines = (highlightStr?: string): number[] => {
   const parts = highlightStr.split(',');
   for (const part of parts) {
     if (part.includes('-')) {
-      const [start, end] = part.split('-').map((n) => Number.parseInt(n.trim(), 10));
+      const [start, end] = part
+        .split('-')
+        .map((n) => Number.parseInt(n.trim(), 10));
       for (let i = start; i <= end; i++) {
         highlightedLineNumbers.push(i);
       }
